requirements.txt
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.108.0
pandas==2.1.3
plotly==5.18.0
python-dotenv==1.0.0 